requirements.txt
discord.py
aiosqlite3
youtube-dlc
pynacl
python-dotenv
requests